Split image viewer stuff into separate modules
