+++ naturtag/qt_app/image_gallery.py
@@ -4,30 +4,22 @@ from os.path import isfile
 from pathlib import Path
 from urllib.parse import unquote, urlparse
 
-from PySide6.QtCore import QSize, Qt, Signal, Slot
-from PySide6.QtGui import QAction, QDropEvent, QKeySequence, QPixmap, QShortcut
-from PySide6.QtWidgets import (
-    QApplication,
-    QFileDialog,
-    QHBoxLayout,
-    QLabel,
-    QMenu,
-    QSizePolicy,
-    QVBoxLayout,
-    QWidget,
-)
+from PySide6.QtCore import Qt, Signal, Slot
+from PySide6.QtGui import QAction, QDropEvent, QPixmap
+from PySide6.QtWidgets import QApplication, QFileDialog, QHBoxLayout, QLabel, QMenu, QVBoxLayout, QWidget
 from qtawesome import icon as fa_icon
 
 from naturtag.constants import IMAGE_FILETYPES, THUMBNAIL_SIZE_DEFAULT
 from naturtag.image_glob import get_images_from_paths
 from naturtag.models import MetaMetadata
+from naturtag.qt_app.image_window import ImageWindow
 from naturtag.qt_app.layouts import FlowLayout
 from naturtag.thumbnails import get_thumbnail
 
 logger = getLogger(__name__)
 
 
-class ImageViewer(QWidget):
+class ImageGallery(QWidget):
     """Container for displaying local image thumbnails & info"""
 
     def __init__(self):
@@ -61,7 +53,7 @@ class ImageViewer(QWidget):
         if not new_images:
             return
 
-        for file_path in new_images:
+        for file_path in sorted(new_images):
             self.load_image(file_path)
 
     def load_image(self, file_path: str):
@@ -236,84 +228,3 @@ class ThumbnailMetaIcons(QLabel):
         icon_label.setPixmap(icon.pixmap(16, 16))
         icon_label.setStyleSheet('background-color: rgba(0, 0, 0, 0);')
         self.icon_layout.addWidget(icon_label)
-
-
-class ImageWindow(QWidget):
-    """Display a full-size image as a separate window
-
-    Keyboard shortcuts: Escape to close window, Left and Right to cycle through images
-    """
-
-    def __init__(self):
-        super().__init__()
-        self.image_paths: list[str] = []
-        self.selected_path = None
-
-        self.image = PixmapLabel()
-        self.image.setFixedSize(QApplication.primaryScreen().availableSize())
-        self.image.setAlignment(Qt.AlignCenter)
-        self.image_layout = QVBoxLayout(self)
-        self.image_layout.addWidget(self.image)
-        self.image_layout.setContentsMargins(0, 0, 0, 0)
-
-        # Keyboard shortcuts
-        shortcut = QShortcut(QKeySequence(Qt.Key_Escape), self)
-        shortcut.activated.connect(self.close)
-        shortcut = QShortcut(QKeySequence(Qt.Key_Right), self)
-        shortcut.activated.connect(self.select_next_image)
-        shortcut = QShortcut(QKeySequence(Qt.Key_Left), self)
-        shortcut.activated.connect(self.select_prev_image)
-
-    def select_image(self, file_path: str, image_paths: list[str]):
-        """Open window to a selected image, and save other available image paths for navigation"""
-        self.selected_path = file_path
-        self.image_paths = image_paths
-        self.image.setPixmap(QPixmap(file_path))
-        self.showFullScreen()
-
-    def select_image_idx(self, idx: int):
-        """Select an image by index, with wraparound"""
-        if idx < 0:
-            idx = len(self.image_paths) - 1
-        elif idx >= len(self.image_paths):
-            idx = 0
-
-        logger.debug(f'Selecting image {idx}: {self.selected_path}')
-        self.selected_path = self.image_paths[idx]
-        self.image.setPixmap(QPixmap(self.selected_path))
-
-    def select_next_image(self):
-        self.select_image_idx(self.image_paths.index(self.selected_path) + 1)
-
-    def select_prev_image(self):
-        self.select_image_idx(self.image_paths.index(self.selected_path) - 1)
-
-
-class PixmapLabel(QLabel):
-    """A QLabel containing a pixmap that preserves its aspect ratio when resizing"""
-
-    def __init__(self, parent: QWidget = None):
-        super().__init__(parent)
-        self.setMinimumSize(1, 1)
-        self.setScaledContents(False)
-        self._pixmap = None
-
-    def setPixmap(self, pixmap: QPixmap):
-        self._pixmap = pixmap
-        super().setPixmap(self.scaledPixmap())
-
-    def heightForWidth(self, width: int) -> int:
-        if self._pixmap:
-            return (self._pixmap.height() * width) / self._pixmap.width()
-        else:
-            return self.height()
-
-    def sizeHint(self) -> QSize:
-        return QSize(self.width(), self.heightForWidth(self.width()))
-
-    def scaledPixmap(self) -> QPixmap:
-        return self._pixmap.scaled(self.size(), Qt.KeepAspectRatio, Qt.SmoothTransformation)
-
-    def resizeEvent(self):
-        if self._pixmap:
-            super().setPixmap(self.scaledPixmap())
